final piece of the refactor
