+++ src/Database/Files.hs
@@ -140,3 +140,6 @@ getAllProcessedFiles :: (MonadUnliftIO m) => Entity User -> m [Entity ProcessedF
 getAllProcessedFiles user = do
   pool <- liftIO getConnectionPool
   runSqlPool (selectList [ProcessedFileUserId ==. entityKey user] [Asc ProcessedFileFilename]) pool
+
+
+
